requirements.txt
langchain
langchain-text-splitters
llama-index
unstructured


fastapi
uvicorn 
qdrant-client
openai
python-dotenv
pydantic
streamlit